Task12

+++ 16-binary_tree_is_perfect.c
@@ -31,15 +31,15 @@ int binary_tree_is_full(const binary_tree_t *tree)
  */
 int binary_tree_balance(const binary_tree_t *tree)
 {
-	size_t left_height = 0, right_height = 0;
+	int left_height = 0, right_height = 0;
 
 	if (tree == NULL)
 		return (0);
 
-	left_height = binary_tree_height(tree->left);
-	right_height = binary_tree_height(tree->right);
+	left_height = (int)binary_tree_height(tree->left);
+	right_height = (int)binary_tree_height(tree->right);
 
-	return ((int)left_height - right_height);
+	return (left_height - right_height);
 }
 
 /**
@@ -72,7 +72,8 @@ int binary_tree_is_perfect(const binary_tree_t *tree)
 	if (tree == NULL)
 		return (0);
 
-	if ((binary_tree_is_full(tree) == 1) && (binary_tree_balance(tree) == 0))
+	if ((binary_tree_is_full(tree) == 1) &&
+	    (binary_tree_balance(tree) == 0))
 		return (1);
 
 	return (0);
